+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - The Sickend and Shaken condition showed a non-descript 'penalty' roll flavor (#854)
 - Fixed not working Fractional Base Bonuses (#856) (thanks mkahvi!)
 - Fixed an issue some users had with missing images (thanks mkahvi!)
+- Wound Thresholds weren't working correctly (#853)
 
 ### Changelog